requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.31.0
click>=8.1.7
rich>=13.5.0
pydantic>=2.4.0